backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-multipart
pdfplumber
python-docx
python-jose[cryptography]
passlib[bcrypt]
openai
python-dotenv
